Extract server to a separate package
